requirements.txt
requests==2.31.0
requests[socks]==2.31.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.5.0
aiofiles==23.2.1
mongomock==4.1.2
pymongo==4.6.0
apprise==1.6.0
prometheus_client==0.19.0
pytest==7.4.3
slowapi==0.1.9